#19: split up InstanceSynchHandler
