+++ at.ac.tuwien.auto.modelua.papyrus.opcua.diagram.transformation/src/at/ac/tuwien/auto/modelua/papyrus/opcua/diagram/transformation/OpcUaDiagramTransformationResources.java
@@ -1,7 +1,5 @@
 package at.ac.tuwien.auto.modelua.papyrus.opcua.diagram.transformation;
 
-import at.ac.tuwien.auto.modelua.papyrus.opcua.diagram.transformation.transformer.SynchHandler;
-
 public class OpcUaDiagramTransformationResources {
 	public static SynchHandler DIAGRAM_SYNCH_HANDLER = Activator.getSynchHandler();
 }
