+++ at.ac.tuwien.auto.modelua.papyrus.opcua.diagram.transformation/src/at/ac/tuwien/auto/modelua/papyrus/opcua/diagram/transformation/listener/DiagramChangeListener.java
@@ -53,7 +53,7 @@ public class DiagramChangeListener implements IPapyrusListener {
 		if(transform)
 		{
 			Element elem = (Element) notifier ;
-			// printNotficationType(notification);
+			printNotficationType(notification);
 			
 			if(notification.getEventType() == Notification.REMOVE || 
 				notification.getEventType() == Notification.REMOVE_MANY)
