+++ at.ac.tuwien.auto.modelua.papyrus.opcua.diagram.transformation/src/at/ac/tuwien/auto/modelua/papyrus/opcua/diagram/transformation/opcuatouml/OpcUaToUmlTransformer.java
@@ -1,18 +1,12 @@
-package at.ac.tuwien.auto.modelua.papyrus.opcua.diagram.transformation.transformer;
+package at.ac.tuwien.auto.modelua.papyrus.opcua.diagram.transformation.opcuatouml;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map.Entry;
 
-import javax.xml.parsers.ParserConfigurationException;
-
-import org.eclipse.core.resources.IWorkspaceRoot;
-import org.eclipse.core.resources.ResourcesPlugin;
-import org.eclipse.core.runtime.IPath;
 import org.eclipse.emf.common.util.EList;
 import org.eclipse.emf.common.util.URI;
-import org.eclipse.emf.ecore.EClass;
 import org.eclipse.emf.ecore.EObject;
 import org.eclipse.emf.ecore.EStructuralFeature;
 import org.eclipse.emf.ecore.impl.DynamicEObjectImpl;
@@ -31,13 +25,11 @@ import org.eclipse.uml2.uml.Generalization;
 import org.eclipse.uml2.uml.Model;
 import org.eclipse.uml2.uml.Package;
 import org.eclipse.uml2.uml.Profile;
-import org.eclipse.uml2.uml.Relationship;
 import org.eclipse.uml2.uml.Stereotype;
 import org.opcfoundation.ua._2011._03.ua.UANodeSet.AliasTable;
 import org.opcfoundation.ua._2011._03.ua.UANodeSet.DataTypeDefinition;
 import org.opcfoundation.ua._2011._03.ua.UANodeSet.DataTypeField;
 import org.opcfoundation.ua._2011._03.ua.UANodeSet.DataTypePurpose;
-import org.opcfoundation.ua._2011._03.ua.UANodeSet.ListOfReferences;
 import org.opcfoundation.ua._2011._03.ua.UANodeSet.LocalizedText;
 import org.opcfoundation.ua._2011._03.ua.UANodeSet.NodeIdAlias;
 import org.opcfoundation.ua._2011._03.ua.UANodeSet.Reference;
@@ -59,2137 +51,51 @@ import org.opcfoundation.ua._2011._03.ua.UANodeSet.UAVariable;
 import org.opcfoundation.ua._2011._03.ua.UANodeSet.UAVariableType;
 import org.opcfoundation.ua._2011._03.ua.UANodeSet.UAView;
 import org.opcfoundation.ua._2011._03.ua.UANodeSet.UriTable;
-import org.opcfoundation.ua._2011._03.ua.UANodeSet.impl.AliasTableImpl;
-import org.opcfoundation.ua._2011._03.ua.UANodeSet.impl.DataTypeDefinitionImpl;
-import org.opcfoundation.ua._2011._03.ua.UANodeSet.impl.DataTypeFieldImpl;
-import org.opcfoundation.ua._2011._03.ua.UANodeSet.impl.ListOfReferencesImpl;
-import org.opcfoundation.ua._2011._03.ua.UANodeSet.impl.ListOfRolePermissionsImpl;
-import org.opcfoundation.ua._2011._03.ua.UANodeSet.impl.LocalizedTextImpl;
-import org.opcfoundation.ua._2011._03.ua.UANodeSet.impl.NodeIdAliasImpl;
-import org.opcfoundation.ua._2011._03.ua.UANodeSet.impl.ReferenceImpl;
 import org.opcfoundation.ua._2011._03.ua.UANodeSet.impl.RolePermissionImpl;
-import org.opcfoundation.ua._2011._03.ua.UANodeSet.impl.StructureTranslationTypeImpl;
-import org.opcfoundation.ua._2011._03.ua.UANodeSet.impl.TranslationTypeImpl;
 import org.opcfoundation.ua._2011._03.ua.UANodeSet.impl.UADataTypeImpl;
-import org.opcfoundation.ua._2011._03.ua.UANodeSet.impl.UAInstanceImpl;
-import org.opcfoundation.ua._2011._03.ua.UANodeSet.impl.UAMethodArgumentImpl;
 import org.opcfoundation.ua._2011._03.ua.UANodeSet.impl.UAMethodImpl;
-import org.opcfoundation.ua._2011._03.ua.UANodeSet.impl.UANodeImpl;
 import org.opcfoundation.ua._2011._03.ua.UANodeSet.impl.UAObjectImpl;
 import org.opcfoundation.ua._2011._03.ua.UANodeSet.impl.UAObjectTypeImpl;
 import org.opcfoundation.ua._2011._03.ua.UANodeSet.impl.UAReferenceTypeImpl;
-import org.opcfoundation.ua._2011._03.ua.UANodeSet.impl.UATypeImpl;
 import org.opcfoundation.ua._2011._03.ua.UANodeSet.impl.UAVariableImpl;
 import org.opcfoundation.ua._2011._03.ua.UANodeSet.impl.UAVariableTypeImpl;
 import org.opcfoundation.ua._2011._03.ua.UANodeSet.impl.UAViewImpl;
 import org.opcfoundation.ua._2011._03.ua.UANodeSet.impl.UriTableImpl;
 
-import at.ac.tuwien.auto.modelua.papyrus.opcua.console.OpcUaMessageConsole;
 import at.ac.tuwien.auto.modelua.papyrus.opcua.libraries.OpcUaLibraryResources;
-import at.ac.tuwien.auto.modelua.papyrus.opcua.nodeset.parser.NodeSetParser;
 import at.ac.tuwien.auto.modelua.papyrus.opcua.preferences.PreferenceProvider;
 
-public class InstanceSyncHandler {
+public class OpcUaToUmlTransformer {
 	
 	private Model baseUmlModel;
 	private UANodeSetType baseNodeset;
 	private HashMap<Element, Object> matching;
 	private HashMap<String, Element> nodeIdMap;
 	private HashMap<String, String> aliasTable;
-	
-	public InstanceSyncHandler(Model umlModel)
-	{
-		this.baseUmlModel = umlModel;
-		this.matching = new HashMap<Element, Object>();
-		this.nodeIdMap = new HashMap<String, Element>();
-		this.aliasTable = new HashMap<String, String>();
-	}
-	
-	public InstanceSyncHandler(UANodeSetType nodeset)
-	{
-		this.baseNodeset = nodeset;
-		this.matching = new HashMap<Element, Object>();
-		this.nodeIdMap = new HashMap<String, Element>();
-		this.aliasTable = new HashMap<String, String>();
-	}
-	
-	public InstanceSyncHandler(Model umlModel, UANodeSetType nodeset)
+		
+	public OpcUaToUmlTransformer(Model umlModel, UANodeSetType nodeset, HashMap<Element, Object> matching, HashMap<String, Element> nodeIdMap)
 	{
 		this.baseUmlModel = umlModel;
 		this.baseNodeset = nodeset;
-		this.matching = new HashMap<Element, Object>();
-		this.nodeIdMap = new HashMap<String, Element>();
+		this.matching = matching;
+		this.nodeIdMap = nodeIdMap;
 		this.aliasTable = new HashMap<String, String>();
 	}
-	
-	public boolean importPackage(String filepath) 
-	{
-		URI fileUri = URI.createFileURI(filepath);
-		ResourceSet owner_resource = this.baseUmlModel.eResource().getResourceSet(); 
-		Model modelimport = (Model) PackageUtil.loadPackage(fileUri, owner_resource);
-				
-		Profile nodeSetProfile = this.baseUmlModel.getAppliedProfile("NodeSet");
-		if(modelimport.isProfileApplied(nodeSetProfile))
-		{
-			// Write Operations have to be executed inside a TransactionalEditingDomain
-			// easiest way to do this is in a Command
-			TransactionalEditingDomain domain = TransactionUtil.getEditingDomain(this.baseUmlModel);
 			
-			// changes to diagrams shall be done inside commands
-			ImportNodeSetCommand cmd = new ImportNodeSetCommand(domain);
-			cmd.setImportModel(modelimport);
-			cmd.setBaseModel(this.baseUmlModel);
-			
-			domain.getCommandStack().execute(cmd);
-			
-			return true;
-		}
-		
-		return false;
-	}
-	
-	public boolean writeToNodeSetFile() throws ParserConfigurationException
-	{
-		IWorkspaceRoot root = ResourcesPlugin.getWorkspace().getRoot();
-		
-        IPath filePath = root.getLocation();
-        URI uri = this.baseUmlModel.eResource().getURI();
-		String path = uri.trimFileExtension().devicePath().substring("/resource/".length())+".xml";
-		filePath=filePath.append(path);
-		return writeToNodeSetFile(filePath.toOSString());
-	}
-	
-	public boolean writeToNodeSetFile(String filepath) throws ParserConfigurationException
-	{
-		if(!filepath.endsWith(".xml"))
-		{
-			filepath = filepath+".xml";
-		}
-		OpcUaMessageConsole.debug("Writing NodeSet file to: "+filepath);
-		return NodeSetParser.writeNodeSetFile(filepath, this.baseNodeset);
-        
-	}
-
-	
-	public void setNodeSetObject(UANodeSetType nodeset)
-	{
-		this.baseNodeset = nodeset;
-	}
-	
-	public void setUmlModel (Model umlModel)
-	{
-		this.baseUmlModel = umlModel;
-	}
-		
-	public boolean transformMember(Element object)
-	{
-		boolean return_val = false;
-		TransactionalEditingDomain domain = TransactionUtil.getEditingDomain(this.baseUmlModel);
-		
-		if(object instanceof Class)
-		{			
-			// changes to diagrams shall be done inside commands
-			UpdateUMLClassCommand cmd = new UpdateUMLClassCommand(domain);
-			cmd.setClass((Class) object);
-			cmd.registerInstanceSynchHandler(this);
-			
-			domain.getCommandStack().execute(cmd);
-			
-			return_val = cmd.getCommandResult();
-		}
-		else if(object instanceof Generalization)
-		{
-			// changes to diagrams shall be done inside commands
-			UpdateUMLGeneralizationCommand cmd = new UpdateUMLGeneralizationCommand(domain);
-			cmd.setGeneralization((Generalization) object);
-			cmd.registerInstanceSynchHandler(this);
-			
-			domain.getCommandStack().execute(cmd);
-			
-			return_val = cmd.getCommandResult();
-		}
-		else if(object instanceof Model)
-		{
-			Profile nodeSetProfile = this.baseUmlModel.getAppliedProfile("NodeSet");
-			Stereotype nodeSetType   = nodeSetProfile.getOwnedStereotype("UANodeSetType");
-			
-			DynamicEObjectImpl stereotype = (DynamicEObjectImpl) object.getStereotypeApplication(nodeSetType);
-			if(stereotype != null)
-			{
-				return_val=transformUANodeSetType((Model) object, stereotype);
-				if(return_val)
-				{
-					return_val &= transformModel((Model) object);
-				}
-			}
-		}
-		else if(object instanceof Package)
-		{
-			
-			// changes to diagrams shall be done inside commands
-			UpdateUMLPackageCommand cmd = new UpdateUMLPackageCommand(domain);
-			cmd.setPackage((Package) object);
-			cmd.registerInstanceSynchHandler(this);
-			
-			domain.getCommandStack().execute(cmd);
-			
-			return_val = cmd.getCommandResult();
-		}
-		
-		return return_val;
-	}
-	
-	private boolean transformModel(Model model) {
-		boolean success = true;
-		
-		for(Element member : model.getOwnedElements())
-		{
-			if(!this.matching.containsKey(member))
-			{
-				success &= transformMember(member);
-			}
-		}
-		
-		return success;
-	}
-
-	protected boolean transformNamespace(Package namespace) {
-		
-		if(namespace.getURI() == null || namespace.getURI().length() == 0)
-		{
-			// no namespace Uir set
-			return false;
-		}
-		
-		// only modify namespace when its inside the same model
-		if(this.baseUmlModel.equals(namespace.getModel()))
-		{			
-			namespace.setName(namespace.getURI());
-		}
-		
-		Profile nodeSetProfile = this.baseUmlModel.getAppliedProfile("NodeSet");
-		Stereotype nodeSetType   = nodeSetProfile.getOwnedStereotype("UANodeSetType");
-		
-		if(this.baseNodeset.getNamespaceUris() == null)
-		{
-			UriTableImpl nsTable = new UriTableImpl();
-			this.baseNodeset.setNamespaceUris(nsTable);
-		}
-		
-		EcoreEList<DynamicEObjectImpl> nsList = (EcoreEList<DynamicEObjectImpl>) this.baseUmlModel.getValue(nodeSetType, "nameSpaceUris");
-		this.baseNodeset.getNamespaceUris().getUri().clear();
-		boolean already_added = false;
-		
-		for(DynamicEObjectImpl nsObject : nsList )
-		{
-			Package ns = (Package) nsObject.dynamicGet(0);
-			
-			if(namespace.equals(ns))
-			{
-				already_added = true;
-			}
-			
-			if(ns.getURI() != null)
-			{
-				this.baseNodeset.getNamespaceUris().getUri().add(ns.getURI());
-			}
-		}
-		
-		if(!already_added)
-		{
-			Stereotype nameSpaceType   = nodeSetProfile.getOwnedStereotype("NameSpace");
-			nsList.add((DynamicEObjectImpl) namespace.getStereotypeApplication(nameSpaceType));
-		}
-		
-		if(!this.baseNodeset.getNamespaceUris().getUri().contains(namespace.getURI()))
-		{
-			this.baseNodeset.getNamespaceUris().getUri().add(namespace.getURI());
-		}
-		
-		return true;
-	}
-
-	protected boolean transformClass(Class object)
-	{
-		EList<EObject> stereotype_applications = object.getStereotypeApplications();
-
-		boolean success=false;
-		
-		for(EObject ster : stereotype_applications)
-		{
-			DynamicEObjectImpl stereotype = (DynamicEObjectImpl) ster;
-			EClass cls = ster.eClass();
-			
-			switch(cls.getName()) {			
-			case "AliasTable":
-				success= transformAliasTable(object, stereotype);
-				break;
-			case "DataTypeDefinition":
-				success= transformDataTypeDefinition(object, stereotype);
-				break;
-			case "DataTypeField":
-				success=transformDataTypeField(object, stereotype);
-				break;
-			case "ExtensionType":
-				success=transformExtensionType(object);
-				break;
-			case "ModelTableEntry":
-				success=transformModelTableEntry(object);
-				break;
-			case "NodeIdAlias":
-				success=transformNodeIdAlias(object, stereotype);
-				break;
-			case "RolePermission":
-				success=transformRolePermission(object);
-				break;
-			case "StructureTranslationType":
-				success=transformStructureTranslationType(object, stereotype);
-				break;
-			case "TranslationType":
-				success=transformTranslationType(object, stereotype);
-				break;
-			case "UADataType":
-				success=transformUADataType(object);
-				break;
-			case "UAMethod":
-				success=transformUAMethod(object);
-				break;
-			case "UAMethodArgument":
-				success=transformUAMethodArgument(object);
-				break;
-			case "UAObject":
-				success=transformUAObject(object, stereotype);
-				break;
-			case "UAObjectType":
-				success=transformUAObjectType(object);
-				break;
-			case "UAReferenceType":
-				success=transformUAReferenceType(object);
-				break;
-			case "UAVariable":
-				success=transformUAVariable(object);
-				break;
-			case "UAVariableType":
-				success=transformUAVariableType(object);
-				break;
-			case "UAView":
-				success=transformUAView(object);
-				break;
-			case "ValueType":
-				success=transformValueType(object);
-				break;
-			case "ValueType1":
-				success=transformValueType1(object);
-				break;
-			default:
-				break;
-			}
-		}
-		
-		return success;
-		
-	}
-	
-	public boolean removeMember(Element obj) {
-		
-		if(!this.matching.containsKey(obj))
-		{
-			return true;
-		}
-		
-		boolean success = false;
-		Object toBeRemoved = this.matching.get(obj);
-		
-		if(toBeRemoved instanceof UADataType)
-		{
-			success = this.baseNodeset.getUADataType().remove(toBeRemoved);
-		}
-		else if(toBeRemoved instanceof UAMethod)
-		{
-			success = this.baseNodeset.getUAMethod().remove(toBeRemoved);
-		}
-		else if(toBeRemoved instanceof UAObjectType)
-		{
-			success = this.baseNodeset.getUAObjectType().remove(toBeRemoved);
-		}
-		else if(toBeRemoved instanceof UAObject)
-		{
-			success = this.baseNodeset.getUAObject().remove(toBeRemoved);
-		}
-		else if(toBeRemoved instanceof UAReferenceType)
-		{
-			success = this.baseNodeset.getUAReferenceType().remove(toBeRemoved);
-		}
-		else if(toBeRemoved instanceof UAVariable)
-		{
-			success = this.baseNodeset.getUAVariable().remove(toBeRemoved);
-		}
-		else if(toBeRemoved instanceof UAVariableType)
-		{
-			success = this.baseNodeset.getUAVariableType().remove(toBeRemoved);
-		}else if(toBeRemoved instanceof Reference)
-		{
-			success = deleteUAReference((Reference) toBeRemoved, (Generalization) obj);
-		}
-		
-		if(success)
-		{
-			this.matching.remove(obj);
-		}
-		
-		return success;
-	}
-	
-	
-	private boolean deleteUAReference(Reference reference, Generalization general) {
-		EList<Element> sources = general.getSources();
-		
-		for(Element source : sources)
-		{
-			UANode node = (UANode) this.matching.get(source);
-			node.getReferences().getReference().remove(reference);
-		}
-		
-		return true;
-	}
-
-	private boolean transformValueType1(Class object) {
-		// TODO: transform ValueType 1
-		return false;
-		
-	}
-
-	private boolean transformValueType(Class object) {
-		// TODO: transform ValueType 
-		return false;
-	}
-
-	private boolean transformUAView(Class object) {
-		
-		UAViewImpl uaView;
-		if(this.matching.containsKey(object))
-		{
-			uaView = (UAViewImpl) this.matching.get(object);
-		}
-		else
-		{
-			uaView = new UAViewImpl();
-			this.matching.put(object, uaView);
-			if(object.getModel().equals(this.baseUmlModel))
-			{
-				this.baseNodeset.getUAView().add(uaView);
-			}
-		}
-		
-		boolean success = transformUAInstance(object);
-		
-		if(success)
-		{
-			Stereotype uaStereotype = getMatchingStereotype(uaView);
-			
-
-			if(object.hasValue(uaStereotype, "containsNoLoops"))
-			{
-				String stringToConvert = String.valueOf(object.getValue(uaStereotype, "containsNoLoops"));
-				boolean convertedBoolean = Boolean.parseBoolean(stringToConvert);
-				uaView.setContainsNoLoops(convertedBoolean);
-			}
-			else
-			{
-				uaView.unsetContainsNoLoops();
-			}
-			
-			if(object.hasValue(uaStereotype, "eventNotifier"))
-			{
-				String stringToConvert = String.valueOf(object.getValue(uaStereotype, "eventNotifier"));
-				short convertedShort = Short.parseShort(stringToConvert);
-				uaView.setEventNotifier(convertedShort);
-			}
-			else
-			{
-				uaView.unsetEventNotifier();
-			}
-			
-		}
-
-		return true;
-		
-	}
-
-	private boolean transformUAVariableType(Class object) {
-		
-		UAVariableTypeImpl uaVarType;
-		if(this.matching.containsKey(object))
-		{
-			uaVarType = (UAVariableTypeImpl) this.matching.get(object);
-		}
-		else
-		{
-			uaVarType = new UAVariableTypeImpl();
-			this.matching.put(object, uaVarType);
-			if(object.getModel().equals(this.baseUmlModel))
-			{
-				this.baseNodeset.getUAVariableType().add(uaVarType);
-			}
-		}
-		
-		return transformUAType(object);
-	}
-
-	private boolean transformUAVariable(Class object) {
-		
-		UAVariableImpl uaVariable;
-		if(this.matching.containsKey(object))
-		{
-			uaVariable = (UAVariableImpl) this.matching.get(object);
-		}
-		else
-		{
-			uaVariable = new UAVariableImpl();
-			this.matching.put(object, uaVariable);
-			if(object.getModel().equals(this.baseUmlModel))
-			{
-				this.baseNodeset.getUAVariable().add(uaVariable);
-			}
-		}
-		
-		boolean success = transformUAInstance(object);
-		
-
-		if(success)
-		{
-			Stereotype uaStereotype = getMatchingStereotype(uaVariable);
-			
-			if(object.hasValue(uaStereotype, "value"))
-			{
-				//TODO: add value
-				Object value = object.getValue(uaStereotype, "value");
-
-			}
-			
-			if(object.hasValue(uaStereotype, "translation"))
-			{
-				EList<DynamicEObjectImpl> translationList = (EList<DynamicEObjectImpl> )object.getValue(uaStereotype, "translation");
-				uaVariable.getTranslation().clear();
-				for(DynamicEObjectImpl translation : translationList)
-				{					
-					Class translationClass = getStereotypeBaseClass(translation, true);
-					if(translationClass != null)
-					{
-						TranslationType tt = (TranslationType) this.matching.get(translationClass);
-						uaVariable.getTranslation().add(tt);
-						
-						EList<Classifier> children = object.getNestedClassifiers();
-						if(!children.contains(translationClass))
-						{
-							children.add(translationClass);
-						}
-					}
-				}
-			}
-			
-			if(object.hasValue(uaStereotype, "accessLevel"))
-			{
-				String stringToConvert = String.valueOf(object.getValue(uaStereotype, "accessLevel"));
-				Long convertedLong = Long.parseLong(stringToConvert);
-				uaVariable.setAccessLevel(convertedLong);
-			}
-			else
-			{
-				uaVariable.unsetAccessLevel();
-			}
-			
-			if(object.hasValue(uaStereotype, "arrayDimensions"))
-			{
-				List<Object> arrayDimensions = (List<Object>) object.getValue(uaStereotype, "arrayDimensions");
-				String convertedString ="";
-				
-				for(Object dimension : arrayDimensions)
-				{
-					convertedString = convertedString + "," + String.valueOf(dimension);
-				}
-								
-				uaVariable.setArrayDimensions(convertedString.substring(1));
-			}
-			else
-			{
-				uaVariable.unsetArrayDimensions();
-			}
-			
-			if(object.hasValue(uaStereotype, "dataType"))
-			{
-				DynamicEObjectImpl dataType = (DynamicEObjectImpl) object.getValue(uaStereotype, "dataType");
-
-				Class parentClass = getStereotypeBaseClass(dataType, true);
-				
-				String nodeId = getNodeId(parentClass);	
-
-				if(nodeId != null && nodeId.length() > 0)
-				{					
-					uaVariable.setDataType(nodeId);
-				}
-
-			}
-			else
-			{
-				uaVariable.unsetDataType();;
-			}
-			
-			if(object.hasValue(uaStereotype, "historizing"))
-			{
-				String stringToConvert = String.valueOf(object.getValue(uaStereotype, "historizing"));
-				boolean convertedBoolean = Boolean.parseBoolean(stringToConvert);
-				uaVariable.setHistorizing(convertedBoolean);
-
-			}
-			else
-			{
-				uaVariable.unsetHistorizing();
-			}
-
-			if(object.hasValue(uaStereotype, "minimumSamplingInterval"))
-			{
-				double intervall = (double ) object.getValue(uaStereotype, "minimumSamplingInterval");
-				uaVariable.setMinimumSamplingInterval(intervall);
-
-			}
-
-			else
-			{
-				uaVariable.unsetMinimumSamplingInterval();
-			}
-			
-			
-			if(object.hasValue(uaStereotype, "userAccessLevel"))
-			{
-				String stringToConvert = String.valueOf(object.getValue(uaStereotype, "userAccessLevel"));
-				Long convertedLong = Long.parseLong(stringToConvert);
-				uaVariable.setUserAccessLevel(convertedLong);
-			}
-			else
-			{
-				uaVariable.unsetUserAccessLevel();
-			}
-			
-			if(object.hasValue(uaStereotype, "valueRank"))
-			{
-				String stringToConvert = String.valueOf(object.getValue(uaStereotype, "valueRank"));
-				int convertedInt = Integer.parseInt(stringToConvert);
-				uaVariable.setValueRank(convertedInt);
-
-			}
-			else
-			{
-				uaVariable.unsetValueRank();
-			}
-		}
-		
-		return success;
-	}
-
-	private boolean transformUAType(Class object) {
-		UAType uaObjType;
-		if(this.matching.containsKey(object))
-		{
-			uaObjType = (UAType) this.matching.get(object);
-		}
-		else
-		{
-			uaObjType = new UATypeImpl();
-			this.matching.put(object, uaObjType);
-		}
-				
-		return transformUANode(object);
-	}
-
-	private boolean transformUAReferenceType(Class object) {
-		
-		UAReferenceTypeImpl uaRefType;
-		if(this.matching.containsKey(object))
-		{
-			uaRefType = (UAReferenceTypeImpl) this.matching.get(object);
-		}
-		else
-		{
-			uaRefType = new UAReferenceTypeImpl();
-			this.matching.put(object, uaRefType);
-			if(object.getModel().equals(this.baseUmlModel))
-			{
-				this.baseNodeset.getUAReferenceType().add(uaRefType);
-			}
-		}
-		
-		boolean success = transformUAType(object);
-		
-		if(success)
-		{
-			Stereotype uaStereotype = getMatchingStereotype(uaRefType);
-			if(object.hasValue(uaStereotype, "inverseName"))
-			{
-				EList<String> inverseNameList = (EList<String>) object.getValue(uaStereotype, "inverseName");
-				uaRefType.getInverseName().clear();
-				
-				for(String inverseName : inverseNameList)
-				{
-					LocalizedTextImpl lt = new LocalizedTextImpl();
-					lt.setValue(inverseName);
-					uaRefType.getInverseName().add(lt);
-				}
-			}
-			
-			if(object.hasValue(uaStereotype, "symmetric"))
-			{
-				boolean symmetric = (boolean) object.getValue(uaStereotype, "symmetric");
-				uaRefType.setSymmetric(symmetric);
-			}
-		}
-
-
-		return success;
-	}
-
-	private boolean transformUAObjectType(Class object) {
-		
-		UAObjectTypeImpl uaObjType;
-		if(this.matching.containsKey(object))
-		{
-			uaObjType = (UAObjectTypeImpl) this.matching.get(object);
-		}
-		else
-		{
-			uaObjType = new UAObjectTypeImpl();
-			this.matching.put(object, uaObjType);
-			if(object.getModel().equals(this.baseUmlModel))
-			{
-				this.baseNodeset.getUAObjectType().add(uaObjType);
-			}
-		}
-		
-		boolean success = transformUAType(object);
-		
-		return success;
-	}
-
-	private boolean transformUAObject(Class object,  DynamicEObjectImpl stereotype) {
-		UAObjectImpl uaObject;
-		if(this.matching.containsKey(object))
-		{
-			uaObject = (UAObjectImpl) this.matching.get(object);
-		}
-		else
-		{
-			uaObject = new UAObjectImpl();
-			this.matching.put(object, uaObject);
-			if(object.getModel().equals(this.baseUmlModel))
-			{
-				this.baseNodeset.getUAObject().add(uaObject);
-			}
-		}
-		
-		boolean success = transformUAInstance(object);
-		
-		if(success)
-		{
-			Stereotype uaStereotype = getMatchingStereotype(uaObject);
-			
-			if(object.hasValue(uaStereotype, "eventNotifier"))
-			{
-				String stringToConvert = String.valueOf(object.getValue(uaStereotype, "eventNotifier"));
-				short convertedShort = Short.parseShort(stringToConvert);
-				uaObject.setEventNotifier(convertedShort);
-			}
-			else
-			{
-				uaObject.unsetEventNotifier();
-			}
-		}
-			
-		return success;
-	}
-
-	private boolean transformUANodeSetType(Model object,  DynamicEObjectImpl stereotype) {
-		EList<EStructuralFeature> featuresList = stereotype.eClass().getEAllStructuralFeatures();
-		
-		Profile nodeSetProfile = this.baseUmlModel.getAppliedProfile("NodeSet");
-		Stereotype nodeSetType   = nodeSetProfile.getOwnedStereotype("UANodeSetType");
-		
-		if(this.baseUmlModel.hasValue(nodeSetType, "nameSpaceUris"))
-		{			
-			EcoreEList<DynamicEObjectImpl> nsList = (EcoreEList<DynamicEObjectImpl>) this.baseUmlModel.getValue(nodeSetType, "nameSpaceUris");
-			
-			for(DynamicEObjectImpl nsObject : nsList )
-			{
-				Package ns = (Package) nsObject.dynamicGet(0);
-				
-				if(ns.getURI() != null)
-				{
-					this.baseNodeset.getNamespaceUris().getUri().add(ns.getURI());
-				}
-			}
-		}
-		
-		if(this.baseUmlModel.hasValue(nodeSetType, "serverUris"))
-		{
-			EcoreEList<String> serverUriList = (EcoreEList<String>) this.baseUmlModel.getValue(nodeSetType, "serverUris");
-			
-			if(this.baseNodeset.getServerUris() == null)
-			{
-				this.baseNodeset.setServerUris(new UriTableImpl());
-			}
-			
-			this.baseNodeset.getServerUris().getUri().clear();
-			
-			for(String serverUri : serverUriList)
-			{
-				this.baseNodeset.getServerUris().getUri().add(serverUri);
-			}
-		}		
-		
-//		if(object.getValue(nodeSetType, "models") != null)
-//		{
-//				EDataTypeUniqueEList<String> models = (EDataTypeUniqueEList<String>) object.getValue(nodeSetType, "models");
-//		}
-//		if(object.getValue(nodeSetType, "aliases") != null)
-//		{
-//				EDataTypeUniqueEList<String> aliases = (EDataTypeUniqueEList<String>) object.getValue(nodeSetType, "aliases");
-//		}
-//		
-//		if(object.getValue(nodeSetType, "extensions") != null)
-//		{
-//			EDataTypeUniqueEList<String> extensions = (EDataTypeUniqueEList<String>) object.getValue(nodeSetType, "extensions");
-//		}
-		
-		
-		for(EStructuralFeature feature : featuresList)
-		{
-			int id = feature.getFeatureID();
-			String name = feature.getName();
-			Object temp = stereotype.dynamicGet(id);
-			
-			if(name.equalsIgnoreCase("models"))
-			{
-				if(temp == null)
-				{
-					continue;
-				}
-			}
-			else if(name.equalsIgnoreCase("aliases"))
-			{
-				if(temp == null)
-				{
-					continue;
-				}
-				
-				DynamicEObjectImpl alias_object = ((DynamicEObjectImpl) temp);
-				EList<EStructuralFeature> featuresList2 = alias_object.eClass().getEAllStructuralFeatures();
-				for(EStructuralFeature feature2 : featuresList2)
-				{
-					int id2 = feature2.getFeatureID();
-					String name2 = feature2.getName();
-					if(name2.equalsIgnoreCase("base_Class"))
-					{
-						Object baseClass = alias_object.dynamicGet(id2);
-						if(this.matching.containsKey(baseClass))
-						{
-							AliasTableImpl aliasTable = (AliasTableImpl) this.matching.get(baseClass);
-							this.baseNodeset.setAliases(aliasTable);
-							break;
-						}
-						else
-						{
-							transformClass((Class) baseClass);
-							AliasTableImpl aliasTable = (AliasTableImpl) this.matching.get(baseClass);
-							this.baseNodeset.setAliases(aliasTable);
-							break;
-						}
-					}
-				}
-			}
-			else if(name.equalsIgnoreCase("extensions"))
-			{
-				if(temp == null)
-				{
-					continue;
-				}
-			}
-		}
-		
-		return true;
-		
-	}
-
-	private boolean transformUANode(Class object) {
-		
-		UANodeImpl uanode;
-		if(this.matching.containsKey(object))
-		{
-			uanode = (UANodeImpl) this.matching.get(object);
-		}
-		else
-		{
-			uanode = new UANodeImpl();
-			this.matching.put(object, uanode);
-		}
-
-		Stereotype uaStereotype = getMatchingStereotype(uanode);
-
-		if(object.hasValue(uaStereotype, "browseName"))
-		{
-			String browseName = (String) object.getValue(uaStereotype, "browseName");
-			object.setName(browseName);
-			
-			Package namespace = object.getNearestPackage();
-			if(namespace != null && namespace.getURI() != null && namespace.getURI().length() > 0)
-			{
-				if(this.baseNodeset.getNamespaceUris() == null)
-				{
-					this.baseNodeset.setNamespaceUris(new UriTableImpl());
-				}
-				
-				EList<String> namespaces = this.baseNodeset.getNamespaceUris().getUri();
-				
-				int namespaceId = namespaces.indexOf(namespace.getURI());
-				namespaceId++; // arrays start at 0
-				browseName = String.valueOf(namespaceId) + ":" +browseName;
-			}
-			
-			uanode.setBrowseName( browseName);		
-		}
-		
-		if(object.hasValue(uaStereotype, "nodeId"))
-		{
-			String nodeId = getNodeId(object);	
-
-			if(nodeId != null && nodeId.length() > 0)
-			{					
-				uanode.setNodeId(nodeId);
-			}
-			this.nodeIdMap.put(nodeId, object);
-		}
-		else
-		{
-			return false;
-		}
-
-		if(object.hasValue(uaStereotype, "displayName"))
-		{
-			EDataTypeUniqueEList<String> displayName = (EDataTypeUniqueEList<String>) object.getValue(uaStereotype, "displayName");
-			uanode.getDisplayName().clear();
-			for(String dnEntry : displayName)
-			{
-				LocalizedTextImpl lt = new LocalizedTextImpl();
-				lt.setValue(dnEntry);
-				uanode.getDisplayName().add(lt);
-			}
-		}
-		else
-		{
-			uanode.getDisplayName().clear();
-		}
-		
-		if(object.hasValue(uaStereotype, "description"))
-		{
-			EDataTypeUniqueEList<String> description = (EDataTypeUniqueEList<String>) object.getValue(uaStereotype, "description");
-			uanode.getDescription().clear();
-			for(String dEntry : description)
-			{
-				LocalizedTextImpl lt = new LocalizedTextImpl();
-				lt.setValue(dEntry);
-				uanode.getDescription().add(lt);
-			}
-		}
-		else
-		{
-			uanode.getDescription().clear();
-		}
-		
-		if(object.hasValue(uaStereotype, "category"))
-		{
-			EDataTypeUniqueEList<String> category = (EDataTypeUniqueEList<String>) object.getValue(uaStereotype, "category");
-			uanode.getCategory().clear();
-			uanode.getCategory().addAll(category);
-		}
-		else
-		{
-			uanode.getCategory().clear();
-		}
-		
-		if(object.hasValue(uaStereotype, "documentation"))
-		{
-			String documentation = (String) object.getValue(uaStereotype, "documentation");
-			uanode.setDocumentation( documentation);
-		}
-		else
-		{
-			uanode.setDocumentation("");
-		}
-		
-		
-		if(object.hasValue(uaStereotype, "rolePermissions"))
-		{
-
-			EList<Classifier> classifierList = object.getNestedClassifiers();
-			EcoreEList<DynamicEObjectImpl> rolePermissions = (EcoreEList<DynamicEObjectImpl>) object.getValue(uaStereotype, "rolePermissions");
-			
-			for(DynamicEObjectImpl rolePermission : rolePermissions)
-			{
-				Class rolePermissionsClass = getStereotypeBaseClass(rolePermission, true);
-				
-				
-				if(!classifierList.contains(rolePermissionsClass))
-				{
-					classifierList.add(rolePermissionsClass);
-				}
-				
-				if(uanode.getRolePermissions() == null)
-				{
-					uanode.setRolePermissions(new ListOfRolePermissionsImpl());
-				}
-				
-				if(this.matching.containsKey(rolePermissionsClass))
-				{
-					RolePermission uaRolePermission = (RolePermission) this.matching.get(rolePermissionsClass);
-					uanode.getRolePermissions().getRolePermission().add(uaRolePermission);
-				}
-			}
-		}
-		else 
-		{
-			if(uanode.getRolePermissions() == null)
-			{
-				uanode.setRolePermissions(new ListOfRolePermissionsImpl());
-			}
-			else
-			{				
-				uanode.getRolePermissions().getRolePermission().clear();
-			}
-		}
-
-
-		
-		if(object.hasValue(uaStereotype, "accessRestrictions"))
-		{
-			Object temp = object.getValue(uaStereotype, "accessRestrictions");
-			String stringToConvert = String.valueOf(temp);
-			short convertedShort = Short.parseShort(stringToConvert);
-			uanode.setAccessRestrictions(convertedShort);
-		}
-		else
-		{
-			uanode.unsetAccessRestrictions();
-		}
-		
-		EnumerationLiteral releaseStatus = (EnumerationLiteral) object.getValue(uaStereotype, "releaseStatus");
-		String value = releaseStatus.getLabel();
-		switch(value)
-		{
-		case "Draft":
-			uanode.setReleaseStatus(ReleaseStatus.DRAFT);
-			break;
-		case "Released":
-			uanode.setReleaseStatus(ReleaseStatus.RELEASED);
-			break;
-		case "Deprecated":
-			uanode.setReleaseStatus(ReleaseStatus.DEPRECATED);
-			break;
-		default:
-			uanode.unsetReleaseStatus();
-		}
-		
-
-		if(object.hasValue(uaStereotype, "symbolicName"))
-		{
-			EDataTypeUniqueEList<String> symbolicName = (EDataTypeUniqueEList<String>) object.getValue(uaStereotype, "symbolicName");
-			uanode.setSymbolicName(symbolicName);
-		}
-		else
-		{
-			if(uanode.getSymbolicName() != null)
-			{				
-				uanode.getSymbolicName().clear();
-			}
-		}
-		
-		if(object.hasValue(uaStereotype, "userWriteMask"))
-		{
-			Object temp = object.getValue(uaStereotype, "userWriteMask");
-			String stringToConvert = String.valueOf(temp);
-			long convertedLong = Long.parseLong(stringToConvert);
-			uanode.setUserWriteMask(convertedLong);
-		}
-		else
-		{
-			uanode.unsetUserWriteMask();
-		}
-		
-		if(object.hasValue(uaStereotype, "writeMask"))
-		{
-			Object temp = object.getValue(uaStereotype, "writeMask");
-			String stringToConvert = String.valueOf(temp);
-			long convertedLong = Long.parseLong(stringToConvert);
-			uanode.setWriteMask(convertedLong);
-		}
-		else
-		{
-			uanode.unsetWriteMask();
-		}
-		
-		ListOfReferences listOfReferences = uanode.getReferences();
-		if(listOfReferences == null)
-		{
-			listOfReferences = new ListOfReferencesImpl();
-			uanode.setReferences(listOfReferences);
-		}
-		
-		listOfReferences.getReference().clear();
-		
-		for(Generalization reference :object.getGeneralizations())
-		{
-			if(! this.matching.containsKey(reference))
-			{
-				transformGeneralization(reference);
-			}
-			listOfReferences.getReference().add((Reference) this.matching.get(reference));
-
-		}
-
-		return true;
-	}
-
-
-
-	private boolean transformUAMethod(Class object) {
-		
-		UAMethodImpl uaMethod;
-		if(this.matching.containsKey(object))
-		{
-			uaMethod = (UAMethodImpl) this.matching.get(object);
-		}
-		else
-		{
-			uaMethod = new UAMethodImpl();
-			this.matching.put(object, uaMethod);
-			if(object.getModel().equals(this.baseUmlModel))
-			{				
-				this.baseNodeset.getUAMethod().add(uaMethod);
-			}
-		}
-		
-		boolean success = transformUAInstance(object);
-		
-		if(success)
-		{
-			Stereotype uaStereotype = getMatchingStereotype(uaMethod);
-
-			if(object.hasValue(uaStereotype, "argumentDescription"))
-			{
-				EcoreEList<DynamicEObjectImpl> argumentDescriptionList = (EcoreEList<DynamicEObjectImpl>) object.getValue(uaStereotype, "argumentDescription");
-				for(DynamicEObjectImpl argumentDescriptionEntry : argumentDescriptionList)
-				{
-					Class argumentDescription = getStereotypeBaseClass(argumentDescriptionEntry, true);
-					if(argumentDescription != null)
-					{
-						UAMethodArgument uaMethArg = (UAMethodArgument) this.matching.get(argumentDescription);
-						if(!uaMethod.getArgumentDescription().contains(uaMethArg))
-						{						
-							uaMethod.getArgumentDescription().add(uaMethArg);
-						}
-						
-						EList<Classifier> children = object.getNestedClassifiers();
-						if(!children.contains(argumentDescription))
-						{
-							children.add(argumentDescription);
-						}
-					
-					}
-				}		
-			}
-			else
-			{
-				uaMethod.getArgumentDescription().clear();
-			}
-			
-			if(object.hasValue(uaStereotype, "executable"))
-			{
-				boolean executable = (boolean) object.getValue(uaStereotype, "executable");
-				uaMethod.setExecutable(executable);
-			}
-			else
-			{
-				uaMethod.unsetExecutable();
-			}
-			
-			if(object.hasValue(uaStereotype, "methodDeclarationId"))
-			{
-				DynamicEObjectImpl methodDeclarationId = (DynamicEObjectImpl) object.getValue(uaStereotype, "methodDeclarationId");
-
-				Class methodDeclarationClass = getStereotypeBaseClass(methodDeclarationId, true);
-				
-				String nodeId = getNodeId(methodDeclarationClass);	
-
-				if(nodeId != null && nodeId.length() > 0)
-				{					
-					uaMethod.setMethodDeclarationId(nodeId);
-				}
-			}
-			else
-			{
-				uaMethod.setMethodDeclarationId("");
-			}
-			
-			if(object.hasValue(uaStereotype, "userExecutable"))
-			{
-				boolean userExecutable = (boolean) object.getValue(uaStereotype, "executable");
-				uaMethod.setUserExecutable(userExecutable);
-			}
-			else
-			{
-				uaMethod.unsetUserExecutable();
-			}
-			
-			return success;
-		}
-
-		
-
-		return true;
-	}
-
-	private boolean transformUAMethodArgument(Class object) {
-	
-		UAMethodArgumentImpl uaMethArg;
-		
-		if(this.matching.containsKey(object))
-		{
-			uaMethArg = (UAMethodArgumentImpl) this.matching.get(object);
-		}
-		else
-		{
-			uaMethArg = new UAMethodArgumentImpl();
-			this.matching.put(object, uaMethArg);
-		}
-		
-		Stereotype uaStereotype = getMatchingStereotype(uaMethArg);
-
-		if(object.hasValue(uaStereotype, "name"))
-		{
-			String name = String.valueOf(object.getValue(uaStereotype, "name"));
-			uaMethArg.setName(name);
-		}
-		else
-		{
-			uaMethArg.setName("");
-		}
-		
-		if(object.hasValue(uaStereotype, "description"))
-		{
-			EDataTypeUniqueEList<String> descriptionList = (EDataTypeUniqueEList<String>) object.getValue(uaStereotype, "description");
-			uaMethArg.getDescription().clear();
-			
-			for(String description : descriptionList)
-			{
-				LocalizedTextImpl lt = new LocalizedTextImpl();
-				lt.setValue(description);
-				uaMethArg.getDescription().add(lt);
-			}
-		}
-		else
-		{
-			uaMethArg.getDescription().clear();
-		}
-		
-		return false;
-	}
-	
-	private boolean transformUAInstance(Class object) {
-		
-		UAInstanceImpl uaInstance;
-		if(this.matching.containsKey(object))
-		{
-			uaInstance = (UAInstanceImpl) this.matching.get(object);
-		}
-		else
-		{
-			uaInstance = new UAInstanceImpl();
-			this.matching.put(object, uaInstance);
-		}
-				
-		boolean success = transformUANode(object);
-		
-		Stereotype uaStereotype = getMatchingStereotype(uaInstance);
-		if(object.hasValue(uaStereotype, "parentNodeId"))
-		{
-			DynamicEObjectImpl parentNode = (DynamicEObjectImpl) object.getValue(uaStereotype, "parentNodeId");
-
-			Class parentClass = getStereotypeBaseClass(parentNode, true);
-			
-			String nodeId = getNodeId(parentClass);	
-
-			if(nodeId != null && nodeId.length() > 0)
-			{					
-				uaInstance.setParentNodeId(nodeId);
-			}
-		}
-		else
-		{
-			uaInstance.setParentNodeId("");
-		}
-			
-		return success;
-	}
-
-	private boolean transformUADataType(Class object) {
-
-		UADataTypeImpl uaDataType;
-		if(this.matching.containsKey(object))
-		{
-			uaDataType = (UADataTypeImpl) this.matching.get(object);
-		}
-		else
-		{
-			uaDataType = new UADataTypeImpl();
-			this.matching.put(object, uaDataType);
-			if(object.getModel().equals(this.baseUmlModel))
-			{				
-				this.baseNodeset.getUADataType().add(uaDataType);
-			}
-		}
-		boolean success = transformUAType(object);
-		
-		if(success)
-		{
-			Stereotype dataTypeSter = getMatchingStereotype(uaDataType);
-			
-			if(object.hasValue(dataTypeSter, "purpose"))
-			{
-				EnumerationLiteral lit = (EnumerationLiteral) object.getValue(dataTypeSter, "purpose");
-				String value = lit.toString();
-				switch(value)
-				{
-				case "Normal":
-					uaDataType.setPurpose(DataTypePurpose.NORMAL);
-					break;
-				case "CodeGenerator":
-					uaDataType.setPurpose(DataTypePurpose.CODE_GENERATOR);
-					break;
-				case "ServicesOnly":
-					uaDataType.setPurpose(DataTypePurpose.SERVICES_ONLY);
-					break;
-				}
-			}
-			else
-			{
-				uaDataType.unsetPurpose();
-			}
-			
-			
-			if(object.hasValue(dataTypeSter, "definition"))
-			{
-				DynamicEObjectImpl definition = (DynamicEObjectImpl)object.getValue(dataTypeSter, "definition");
-				Class dataTypeDefinition = getStereotypeBaseClass(definition, true);
-				if(dataTypeDefinition != null)
-				{
-					DataTypeDefinition dtd = (DataTypeDefinition) this.matching.get(dataTypeDefinition);
-					uaDataType.setDefinition(dtd);
-					
-					EList<Classifier> children = object.getNestedClassifiers();
-					if(!children.contains(definition))
-					{
-						children.add(dataTypeDefinition);
-					}
-				}
-			}
-			else
-			{
-				uaDataType.setDefinition(null);
-			}
-		}
-
-		return success;
-	}
-
-	private boolean transformTranslationType(Class object,  DynamicEObjectImpl stereotype) {
-		EList<EStructuralFeature> featuresList = stereotype.eClass().getEAllStructuralFeatures();
-		
-		TranslationTypeImpl tt;
-		if( this.matching.containsKey(object))
-		{
-			tt = (TranslationTypeImpl) this.matching.get(object);
-
-		}
-		else
-		{
-			tt = new TranslationTypeImpl();
-			this.matching.put(object, tt);
-		}
-		
-		for(EStructuralFeature feature : featuresList)
-		{
-			int id = feature.getFeatureID();
-			String name = feature.getName();
-			Object temp = stereotype.dynamicGet(id);
-			
-			if(name.equalsIgnoreCase("text"))
-			{
-				tt.getText().clear();
-				List<String> texts = (List<String>) temp;
-				for(String txt : texts)
-				{
-					LocalizedTextImpl lt = new LocalizedTextImpl();
-					lt.setValue(txt);
-					tt.getText().add(lt);
-				}
-			
-			}
-			else if(name.equalsIgnoreCase("field"))
-			{
-				tt.getField().clear();
-				
-				EcoreEList<DynamicEObjectImpl> ltsUML = (EcoreEList<DynamicEObjectImpl>) temp;
-				for(DynamicEObjectImpl ltUML : ltsUML)
-				{
-					EList<EStructuralFeature> featuresList2 = ltUML.eClass().getEAllStructuralFeatures();
-					for(EStructuralFeature feature2 : featuresList2)
-					{
-						int id2 = feature2.getFeatureID();
-						String name2 = feature2.getName();
-						if(name2.equalsIgnoreCase("base_Class"))
-						{
-							Object baseClass = ltUML.dynamicGet(id2);
-							if(this.matching.containsKey(baseClass))
-							{
-								StructureTranslationType existingLT = (StructureTranslationType) this.matching.get(baseClass);
-								tt.getField().add(existingLT);
-								break;
-							}
-							else
-							{
-								transformClass((Class) baseClass);
-								StructureTranslationType existingLT = (StructureTranslationType) this.matching.get(baseClass);
-								tt.getField().add(existingLT);
-								break;
-							}
-						}
-					}
-				}
-			}
-		}
-
-		return true;
-	}
-
-	private boolean transformStructureTranslationType(Class object,  DynamicEObjectImpl stereotype) {
-		EList<EStructuralFeature> featuresList = stereotype.eClass().getEAllStructuralFeatures();
-		
-		StructureTranslationTypeImpl stt;
-		if( this.matching.containsKey(object))
-		{
-			stt = (StructureTranslationTypeImpl) this.matching.get(object);
-
-		}
-		else
-		{
-			stt = new StructureTranslationTypeImpl();
-			this.matching.put(object, stt);
-		}
-		
-		for(EStructuralFeature feature : featuresList)
-		{
-			int id = feature.getFeatureID();
-			String name = feature.getName();
-			Object temp = stereotype.dynamicGet(id);
-			
-			if(name.equalsIgnoreCase("text"))
-			{
-				stt.getText().clear();
-				List<String> texts = (List<String>) temp;
-				for(String txt : texts)
-				{
-					LocalizedTextImpl lt = new LocalizedTextImpl();
-					lt.setValue(txt);
-					stt.getText().add(lt);
-				}
-				
-			}
-			else if(name.equalsIgnoreCase("name"))
-			{
-				stt.setName((String) temp);
-			}
-		}
-		
-
-		return true;
-	}
-
-	private boolean transformRolePermission(Class object) {
-		
-		RolePermissionImpl rp;
-		if( this.matching.containsKey(object))
-		{
-			rp = (RolePermissionImpl) this.matching.get(object);
-
-		}
-		else
-		{
-			rp = new RolePermissionImpl();
-			this.matching.put(object, rp);
-		}
-
-		Stereotype uaStereotype = getMatchingStereotype(rp);
-		
-		if(object.hasValue( uaStereotype, "value"))
-		{
-			DynamicEObjectImpl dataType = (DynamicEObjectImpl) object.getValue(uaStereotype, "value");
-
-			Class parentClass = getStereotypeBaseClass(dataType, true);
-			
-			String nodeId = getNodeId(parentClass);	
-
-			if(nodeId != null && nodeId.length() > 0)
-			{					
-				rp.setValue(nodeId);
-			}
-		}
-		else
-		{
-			rp.setValue("");
-		}
-		
-		if(object.hasValue( uaStereotype, "permissions"))
-		{
-			String stringToConvert = String.valueOf(object.getValue( uaStereotype, "permissions"));
-	        Long convertedLong = Long.parseLong(stringToConvert);
-			rp.setPermissions(convertedLong);
-		}
-		else
-		{
-			rp.unsetPermissions();
-		}
-				
-		return true;
-		
-	}
-
-	private boolean transformModelTableEntry(Class object) {
-		//TODO: Implement Model Table transformation 
-		
-		return false;
-	}
-
-
-	private boolean transformExtensionType(Class object) {
-		//TODO: Implement Extension Type transformation 
-
-		return false;
-	}
-	
-	private boolean transformDataTypeField(Class object,  DynamicEObjectImpl stereotype) {
-	
-		DataTypeFieldImpl dtf;
-		if( this.matching.containsKey(object))
-		{
-			dtf = (DataTypeFieldImpl) this.matching.get(object);
-
-		}
-		else
-		{
-			dtf = new DataTypeFieldImpl();
-			this.matching.put(object, dtf);
-		}
-		Stereotype sterDTF = getMatchingStereotype(dtf);
-		
-		if(object.getValue(sterDTF, "displayName") !=null)
-		{			
-
-			EDataTypeUniqueEList<String> displayNameList = (EDataTypeUniqueEList<String>) object.getValue(sterDTF, "displayName");
-			dtf.getDisplayName().clear();
-			
-			for(String displayName : displayNameList)
-			{
-				LocalizedTextImpl lt = new LocalizedTextImpl();
-				lt.setValue(displayName);
-				dtf.getDisplayName().add(lt);
-			}
-		}
-		else
-		{
-			dtf.getDisplayName().clear();
-		}
-		
-		if(object.getValue(sterDTF, "description") !=null)
-		{			
-
-			EDataTypeUniqueEList<String> descriptionList = (EDataTypeUniqueEList<String>) object.getValue(sterDTF, "description");
-			dtf.getDescription().clear();
-			
-			for(String description : descriptionList)
-			{
-				LocalizedTextImpl lt = new LocalizedTextImpl();
-				lt.setValue(description);
-				dtf.getDescription().add(lt);
-			}
-		}
-		else
-		{
-			dtf.getDescription().clear();
-		}
-		
-		if(object.getValue(sterDTF, "abstractDataType") !=null)
-		{			
-			DynamicEObjectImpl abstractDataType = (DynamicEObjectImpl) object.getValue(sterDTF, "abstractDataType");
-			
-			Class umlUaNode = getStereotypeBaseClass(abstractDataType, true);
-			
-			if(!this.matching.containsKey(umlUaNode))
-			{
-				transformClass(umlUaNode);
-			}
-			
-			if(umlUaNode != null)
-			{
-				String nodeId = getNodeId(umlUaNode);	
-
-				if(nodeId != null && nodeId.length() > 0)
-				{					
-					dtf.setAbstractDataType(nodeId);
-				}
-				
-			}
-		}
-		else
-		{
-			dtf.setAbstractDataType("");
-		}
-		
-		if(object.getValue(sterDTF, "arrayDimensions") !=null)
-		{		
-			List<Object> arrayDimensions = (List<Object>) object.getValue(sterDTF, "arrayDimensions");
-			String convertedString ="";
-			
-			for(Object dimension : arrayDimensions)
-			{
-				convertedString = convertedString + "," + String.valueOf(dimension);
-			}
-							
-			dtf.setArrayDimensions(convertedString.substring(1));
-			
-		}
-		else
-		{
-			dtf.unsetArrayDimensions();
-		}
-		
-		if(object.getValue(sterDTF, "dataType") !=null)
-		{			
-			
-			DynamicEObjectImpl abstractDataType = (DynamicEObjectImpl) object.getValue(sterDTF, "dataType");
-			
-			Class umlUaNode = getStereotypeBaseClass(abstractDataType, true);
-			
-			if(!this.matching.containsKey(umlUaNode))
-			{
-				transformClass(umlUaNode);
-			}
-			
-			if(umlUaNode != null)
-			{
-				Stereotype uaStereotype = getMatchingStereotype(this.matching.get(umlUaNode));
-				
-				if(umlUaNode.hasValue(uaStereotype, "nodeId"))
-				{
-					String nodeId = getNodeId(umlUaNode);			
-					if(nodeId != null && nodeId.length() > 0)
-					{					
-						dtf.setDataType(nodeId);
-					}
-					
-				}
-				
-			}
-		}
-		else
-		{
-			dtf.unsetDataType();
-		}
-		
-		if(object.getValue(sterDTF, "isOptional") !=null)
-		{			
-			String stringToConvert = String.valueOf(object.getValue(sterDTF, "isOptional"));
-			boolean convertedBoolean = Boolean.parseBoolean(stringToConvert);
-			dtf.setIsOptional(convertedBoolean);
-		}
-		else
-		{
-			dtf.unsetIsOptional();
-		}
-
-		if(object.getValue(sterDTF, "maxStringLength") !=null)
-		{			
-			String stringToConvert = String.valueOf(object.getValue(sterDTF, "maxStringLength"));
-			long convertedLong = Long.parseLong(stringToConvert);
-			dtf.setMaxStringLength(convertedLong);
-		}
-		else
-		{
-			dtf.unsetMaxStringLength();
-		}
-		
-		if(object.getValue(sterDTF, "name") !=null)
-		{			
-			String stringName = String.valueOf(object.getValue(sterDTF, "name"));
-			dtf.setName(stringName);
-		}
-		else
-		{
-			dtf.setName("");
-		}
-		
-		if(object.getValue(sterDTF, "symbolicName") !=null)
-		{			
-
-			EDataTypeUniqueEList<String> symbolicNameList = (EDataTypeUniqueEList<String>) object.getValue(sterDTF, "symbolicName");
-			if(dtf.getSymbolicName() == null)
-			{
-				dtf.setSymbolicName(new ArrayList<String>());
-			}
-			dtf.getSymbolicName().clear();
-			
-			for(String symbolicName : symbolicNameList)
-			{
-				dtf.getSymbolicName().add(symbolicName);
-			}
-		}
-		else
-		{
-			dtf.getSymbolicName().clear();
-		}
-		
-		if(object.getValue(sterDTF, "value") !=null)
-		{			
-			String stringToConvert = String.valueOf(object.getValue(sterDTF, "value"));
-			int convertedInt = Integer.parseInt(stringToConvert);
-			dtf.setValue(convertedInt);
-		}
-		else
-		{
-			dtf.unsetValue();
-		}
-		
-		if(object.getValue(sterDTF, "valueRank") !=null)
-		{			
-			String stringToConvert = String.valueOf(object.getValue(sterDTF, "valueRank"));
-			int convertedInt = Integer.parseInt(stringToConvert);
-			dtf.setValueRank(convertedInt);
-		}
-		else
-		{
-			dtf.unsetValueRank();
-		}
-		
-		return true;
-		
-	}
-
-	private boolean transformDataTypeDefinition(Class object,  DynamicEObjectImpl stereotype) {
-			
-		DataTypeDefinitionImpl dtd;
-		if( this.matching.containsKey(object))
-		{
-			dtd = (DataTypeDefinitionImpl) this.matching.get(object);
-
-		}
-		else
-		{
-			dtd = new DataTypeDefinitionImpl();
-			this.matching.put(object, dtd);
-		}
-		Stereotype sterDTD = getMatchingStereotype(dtd);
-		
-		if(object.getValue(sterDTD, "field") !=null)
-		{			
-			EcoreEList<DynamicEObjectImpl> field = (EcoreEList<DynamicEObjectImpl>) object.getValue(sterDTD, "field");
-			for(DynamicEObjectImpl fieldEntry : field)
-			{
-				Class dataTypeField = getStereotypeBaseClass(fieldEntry, true);
-				if(dataTypeField != null)
-				{
-					DataTypeField dtf = (DataTypeField) this.matching.get(dataTypeField);
-					if(!dtd.getField().contains(dtf))
-					{						
-						dtd.getField().add(dtf);
-					}
-					
-					EList<Classifier> children = object.getNestedClassifiers();
-					if(!children.contains(dataTypeField))
-					{
-						children.add(dataTypeField);
-					}
-				
-				}
-			}			
-		}
-		else
-		{
-			dtd.getField().clear();
-		}
-		
-		if(object.getValue(sterDTD, "isOptionSet") !=null)
-		{			
-			String stringToConvert = String.valueOf(object.getValue(sterDTD, "isOptionSet"));
-			boolean convertedBoolean = Boolean.parseBoolean(stringToConvert);
-			dtd.setIsOptionSet(convertedBoolean);
-		}
-		else
-		{
-			dtd.unsetIsOptionSet();
-		}
-
-		if(object.getValue(sterDTD, "isUnion") !=null)
-		{			
-			String stringToConvert = String.valueOf(object.getValue(sterDTD, "isUnion"));
-			boolean convertedBoolean = Boolean.parseBoolean(stringToConvert);
-			dtd.setIsUnion(convertedBoolean);
-		}
-		else
-		{
-			dtd.unsetIsUnion();
-		}
-		
-		if(object.getValue(sterDTD, "name") !=null)
-		{			
-			String stringToConvert = String.valueOf(object.getValue(sterDTD, "name"));
-			dtd.setName(stringToConvert);
-		}
-		else
-		{
-			dtd.setName("");
-		}
-		
-		if(object.getValue(sterDTD, "symbolicName") !=null)
-		{			
-
-			EDataTypeUniqueEList<String> symbolicNameList = (EDataTypeUniqueEList<String>) object.getValue(sterDTD, "symbolicName");
-			dtd.getSymbolicName().clear();
-			
-			for(String symbolicName : symbolicNameList)
-			{
-				dtd.getSymbolicName().add(symbolicName);
-			}
-			
-		}
-		else
-		{
-			dtd.unsetSymbolicName();
-		}
-		
-		return true;
-	}
-	
-	private boolean transformAliasTable(Class object, DynamicEObjectImpl stereoptype_application) {
-		EList<EStructuralFeature> featuresList = stereoptype_application.eClass().getEAllStructuralFeatures();		
-		AliasTable aliasTable = this.baseNodeset.getAliases();
-		if(aliasTable == null)
-		{
-			aliasTable = new AliasTableImpl();
-			this.baseNodeset.setAliases(aliasTable);
-			this.matching.put(object, aliasTable);
-		}	
-		aliasTable.getAlias().clear();
-		
-		
-		for(EStructuralFeature feature : featuresList)
-		{
-			int id = feature.getFeatureID();
-			String name = feature.getName();
-			
-			if(name.equalsIgnoreCase("alias"))
-			{
-				EcoreEList<DynamicEObjectImpl> aliasesUML = (EcoreEList<DynamicEObjectImpl>) stereoptype_application.dynamicGet(id);
-				for(DynamicEObjectImpl aliasUML : aliasesUML)
-				{
-					//transformNodeIdAlias(aliasUML);
-					EList<EStructuralFeature> featuresList2 = aliasUML.eClass().getEAllStructuralFeatures();
-					for(EStructuralFeature feature2 : featuresList2)
-					{
-						int id2 = feature2.getFeatureID();
-						String name2 = feature2.getName();
-						if(name2.equalsIgnoreCase("base_Class"))
-						{
-							Object baseClass = aliasUML.dynamicGet(id2);
-							if(this.matching.containsKey(baseClass))
-							{
-								NodeIdAlias existingAlias = (NodeIdAlias) this.matching.get(baseClass);
-								aliasTable.getAlias().add(existingAlias);
-								break;
-							}
-							else
-							{
-								transformClass((Class) baseClass);
-								NodeIdAlias existingAlias = (NodeIdAlias) this.matching.get(baseClass);
-								aliasTable.getAlias().add(existingAlias);
-								break;
-							}
-						}
-					}
-				}
-			}
-		}
-		return true;
-	}
-	
-	private boolean transformNodeIdAlias(Class object, DynamicEObjectImpl stereoptype_application) {
-		
-		EList<EStructuralFeature> featuresList = stereoptype_application.eClass().getEAllStructuralFeatures();
-		NodeIdAliasImpl alias = new NodeIdAliasImpl();
-		
-		for(EStructuralFeature feature : featuresList)
-		{
-			int id = feature.getFeatureID();
-			String name = feature.getName();
-			
-			if(name.equalsIgnoreCase("alias"))
-			{
-				String aliasString = (String) stereoptype_application.dynamicGet(id);
-				alias.setAlias(aliasString);
-			}
-			else if(name.equalsIgnoreCase("value"))
-			{
-				String nodeIdString = (String) stereoptype_application.dynamicGet(id);
-				alias.setValue(nodeIdString);
-			}
-		}
-		
-		if(this.matching.containsKey(object))
-		{
-			NodeIdAlias existingAlias = (NodeIdAlias) this.matching.get(object);
-			existingAlias.setAlias(alias.getAlias());
-			existingAlias.setValue(alias.getValue());	
-		}
-		else
-		{
-			this.matching.put(object, alias);
-		}
-		
-		return true;
-	}
-	
-	protected boolean transformGeneralization(Generalization general) {
-
-		if(general.getSources().size() != 1 || general.getTargets().size() != 1)
-		{
-			// only single source/target is supported
-			return false;
-		}
-		
-		Profile nodeSetProfile = this.baseUmlModel.getAppliedProfile("NodeSet");
-		Stereotype uaStereoType  = nodeSetProfile.getOwnedStereotype("Reference");
-		
-		ReferenceImpl uaReference;
-		
-		if(this.matching.containsKey(general))
-		{
-			uaReference = (ReferenceImpl) this.matching.get(general);
-		}
-		else
-		{
-			uaReference = new ReferenceImpl();
-			this.matching.put(general, uaReference);
-			if(general.hasValue(uaStereoType,"isForward"))
-			{
-				uaReference.setIsForward((boolean) general.getValue(uaStereoType,"isForward"));
-			}
-		}
-		
-		Class target = (Class) general.getTargets().get(0);
-		Class source = (Class) general.getSources().get(0);
-
-		String nodeId = getNodeId(target);
-		uaReference.setValue(nodeId);
-		
-		boolean isForward = uaReference.isIsForward();
-		boolean isForwardOld = isForward;
-		//if(general.hasValue(uaStereoType,"isForward"))
-		{
-			isForward =(boolean) general.getValue(uaStereoType,"isForward");
-		}
-//		isForward =(boolean) general.getValue(uaStereoType,"isForward");
-		
-		boolean directionChanged = isForwardOld ^ isForward;
-		uaReference.setIsForward(isForward);
-
-		if(general.hasValue(uaStereoType,"referenceType"))
-		{
-			Stereotype uaReferenceType  = nodeSetProfile.getOwnedStereotype("UAReferenceType");
-			
-			DynamicEObjectImpl referenceTypeSterAppl = (DynamicEObjectImpl) general.getValue(uaStereoType,"referenceType");
-			Class referenceType = getStereotypeBaseClass(referenceTypeSterAppl, true);
-			
-			String referenceTypeNode = getNodeId(referenceType);
-			uaReference.setReferenceType(referenceTypeNode);
-			
-			boolean isHierachicalReference = (boolean) referenceType.getValue(uaReferenceType, "isHierachical");
-			
-			Package sourcePackage = source.getNearestPackage();
-			String sourceNs = sourcePackage.getURI();
-			if(sourceNs == null)
-			{
-				sourceNs = "";
-			}
-			
-			Package targetPackage = target.getNearestPackage();
-			String targetNs = targetPackage.getURI();
-			if(targetNs == null)
-			{
-				targetNs = "";
-			}
-			
-			general.setValue(uaStereoType,"referenceType_symmetric", referenceType.getValue(uaReferenceType, "symmetric"));
-			general.setValue(uaStereoType,"referenceType_browseName", referenceType.getValue(uaReferenceType, "browseName"));
-			general.setValue(uaStereoType,"referenceType_isHierachical", isHierachicalReference);
-			
-			if(isHierachicalReference && sourceNs.equals(targetNs))
-			{
-				//directionChanged = false;
-				if(isForward)
-				{
-					if(directionChanged || target.getNestedClassifiers().contains(source))
-					{
-						Element owner = target.getOwner();
-						if(owner instanceof Class)
-						{
-							Class ownerClass = (Class)owner;
-							ownerClass.getNestedClassifiers().add(source);
-						}
-						else
-						{
-							// owner is package of Model
-							Package ownerPackage = (Package) owner;
-							target.getNestedClassifiers().remove(source);
-							source.setPackage(ownerPackage);
-						}
-						source.getNestedClassifiers().add(target);
-					}
-					else
-					{
-						EList<Classifier> children = source.getNestedClassifiers();
-						if(!children.contains(target))
-						{
-							children.add(target);
-						}
-					}
-				}
-				else
-				{
-					if(directionChanged || source.getNestedClassifiers().contains(target))
-					{
-						Element owner = source.getOwner();
-						if(owner instanceof Class)
-						{
-							Class ownerClass = (Class)owner;
-							ownerClass.getNestedClassifiers().add(target);
-						}
-						else
-						{
-							// owner is package of Model
-							Package ownerPackage = (Package) owner;
-							source.getNestedClassifiers().remove(target);
-							target.setPackage(ownerPackage);
-						}
-						target.getNestedClassifiers().add(source);
-					}
-					else
-					{						
-						EList<Classifier> children = target.getNestedClassifiers();
-						
-						if(!children.contains(source))
-						{
-							children.add(source);
-						}
-					}
-				}
-			}
-			
-			if(source.isStereotypeApplied(uaReferenceType) && target.isStereotypeApplied(uaReferenceType))
-			{
-				if(isForward)
-				{
-					if(this.baseUmlModel.equals(target.getModel()))
-					{
-						boolean isHierachicalReferenceType = (boolean) source.getValue(uaReferenceType, "isHierachical");
-						target.setValue(uaReferenceType, "isHierachical", isHierachicalReferenceType);
-					}
-				}
-				else
-				{
-					if(this.baseUmlModel.equals(source.getModel()))
-					{
-						boolean isHierachicalReferenceType = (boolean) target.getValue(uaReferenceType, "isHierachical");
-						source.setValue(uaReferenceType, "isHierachical", isHierachicalReferenceType);
-					}
-				}
-				
-			}
-			
-		}
-		
-		if(!this.matching.containsKey(source))
-		{
-			transformClass((Class) source);
-		}
-		
-		UANode src = (UANode) this.matching.get(source);
-		ListOfReferences refList = src.getReferences();
-		if(refList == null)
-		{
-			src.setReferences(new ListOfReferencesImpl());
-		}
-		boolean contains = refList.getReference().contains(uaReference);
-		if(!contains)
-		{
-			refList.getReference().add(uaReference);
-		}
-		
-		return true;
-	}
-	
-	private String getNodeId(Element object)
-	{
-		EList<EObject> applied_stereotypes = object.getStereotypeApplications();
-		for(EObject applied_stereotype : applied_stereotypes)
-		{
-			DynamicEObjectImpl dyn_ster = (DynamicEObjectImpl) applied_stereotype;
-			
-			EList<EStructuralFeature> featureList = dyn_ster.eClass().getEAllStructuralFeatures();
-			for(EStructuralFeature feature : featureList)
-			{
-				String name = feature.getName();
-				if(name.equalsIgnoreCase("nodeId"))
-				{
-					int id = feature.getFeatureID();
-					return (String) dyn_ster.dynamicGet(id);
-				}
-			}
-		}
-		
-		return "";
-	}
-	
-	public boolean updatedMember(UANodeSetType nodeset)
+	public boolean transformNodeSet(UANodeSetType nodeset)
 	{
 		
 		TransactionalEditingDomain domain = TransactionUtil.getEditingDomain(this.baseUmlModel);
 		
 		// changes to diagrams shall be done inside commands
-		UpdateUMLCommand cmd = new UpdateUMLCommand(domain);
-		cmd.setNodeSet(nodeset);
-		cmd.registerInstanceSynchHandler(this);
+		TransformOpcUaCommand cmd = new TransformOpcUaCommand(domain, this, nodeset);
 		
 		domain.getCommandStack().execute(cmd);
 		
 		return cmd.getCommandResult();
 	}
 	
-	protected boolean updateMember(UANodeSetType nodeset)
+	protected boolean transformMember(UANodeSetType nodeset)
 	{
         // Implement your write operations here,
     	boolean success = true;
@@ -2208,9 +114,6 @@ public class InstanceSyncHandler {
     	ArrayList<UADataType> dataTypeDefinitions = new ArrayList<UADataType>();
     	ArrayList<UAReferenceType> referenceTypes = new ArrayList<UAReferenceType>();
     	
-    	// Delete all nodes which are not part of the loaded NodeSet
-    	success &= removeMissingElements(nodeset);
-    	
     	if(nodeset.getUAObjectType() != null)
     	{    		
     		// adding and removing needs to be done via list otherwise 
@@ -2218,6 +121,7 @@ public class InstanceSyncHandler {
     		EList<UAObjectType> uaObjectTypes = nodeset.getUAObjectType();
     		List<UAObjectType> nodesToAdd = new ArrayList<UAObjectType>();
     		List<UAObjectType> nodesToDelete = new ArrayList<UAObjectType>();
+
     		for(UAObjectType t : uaObjectTypes)
     		{
     			if(t.getReferences() != null && t.getReferences().getReference().size() > 0 )
@@ -2237,7 +141,7 @@ public class InstanceSyncHandler {
     				break;
     			}
     		}
-       		
+    		
     		// Important first remove old elements than add new ones
     		this.baseNodeset.getUAObjectType().removeAll(nodesToDelete);
     		this.baseNodeset.getUAObjectType().addAll(nodesToAdd);
@@ -2531,241 +435,9 @@ public class InstanceSyncHandler {
     	    	
 		return success;
 	}
-	
-	private boolean removeMissingElements(UANodeSetType nodeset) {
-		
-		if(this.baseNodeset.getUAObjectType() != null)
-    	{    		
-			ArrayList<String> nodeIds = new ArrayList<String>();
 
-    		EList<UAObjectType> uaObjectTypes = nodeset.getUAObjectType();    		
-    		if(uaObjectTypes != null)
-    		{
-	    		for(UAObjectType type : uaObjectTypes)
-	    		{
-	    			nodeIds.add(type.getNodeId());
-	    		}
-    		}
-    		
-    		uaObjectTypes = this.baseNodeset.getUAObjectType();
-    		ArrayList<UAObjectType> toRemove = new ArrayList<UAObjectType>();
-    		
-    		for(UAObjectType type : uaObjectTypes)
-    		{
-    			if(!nodeIds.contains(type.getNodeId()))
-    			{
-    				destroyMember(type);
-    				toRemove.add(type);
-    			}
-    		}
-    		this.baseNodeset.getUAObjectType().removeAll(toRemove);
-    	}
-    	
-    	if(this.baseNodeset.getUAVariableType() != null)
-    	{    		
-    		ArrayList<String> nodeIds = new ArrayList<String>();
-    		
-    		EList<UAVariableType> uaVariableTypes = nodeset.getUAVariableType();
-    		if(uaVariableTypes != null)
-    		{    			
-    			for(UAVariableType type : uaVariableTypes)
-    			{
-    				nodeIds.add(type.getNodeId());
-    			}
-    		}
-       		
-    		uaVariableTypes = this.baseNodeset.getUAVariableType();
-    		ArrayList<UAVariableType> toRemove = new ArrayList<UAVariableType>();
-    		for(UAVariableType type : uaVariableTypes)
-    		{
-    			if(!nodeIds.contains(type.getNodeId()))
-    			{
-    				destroyMember(type);
-    				toRemove.add(type);
-    			}
-    		}
-    		this.baseNodeset.getUAVariableType().removeAll(toRemove);
-    	}
-    	
-    	if(this.baseNodeset.getUADataType() != null)
-    	{    		
-    		ArrayList<String> nodeIds = new ArrayList<String>();
 
-    		EList<UADataType> uaDataTypes = nodeset.getUADataType();
-    		if(uaDataTypes != null)
-    		{    			
-    			for(UADataType type : uaDataTypes)
-    			{
-    				nodeIds.add(type.getNodeId());
-    			}
-    		}
-       		
-    		uaDataTypes = this.baseNodeset.getUADataType();
-    		ArrayList<UADataType> toRemove = new ArrayList<UADataType>();
-    		for(UADataType type : uaDataTypes)
-    		{
-    			if(!nodeIds.contains(type.getNodeId()))
-    			{
-    				destroyMember(type);
-    				toRemove.add(type);
-    			}
-    		}
-    		this.baseNodeset.getUADataType().removeAll(toRemove);
-    	}
-		
-    	if(this.baseNodeset.getUAReferenceType() != null)
-    	{    		
-    		ArrayList<String> nodeIds = new ArrayList<String>();
-    		
-    		EList<UAReferenceType> uaReferenceTypes = nodeset.getUAReferenceType();
-    		if(uaReferenceTypes != null)
-    		{    			
-    			for(UAReferenceType type : uaReferenceTypes)
-    			{
-    				nodeIds.add(type.getNodeId());
-    			}
-    		}
-       		
-    		uaReferenceTypes = this.baseNodeset.getUAReferenceType();
-    		ArrayList<UAReferenceType> toRemove = new ArrayList<UAReferenceType>();
-    		for(UAReferenceType type : uaReferenceTypes)
-    		{
-    			if(!nodeIds.contains(type.getNodeId()))
-    			{
-    				destroyMember(type);
-    				toRemove.add(type);
-    			}
-    		}
-    		this.baseNodeset.getUAReferenceType().removeAll(toRemove);
-    	}
-    	
-    	if(this.baseNodeset.getUAObject() != null)
-    	{
-    		ArrayList<String> nodeIds = new ArrayList<String>();
-
-			EList<UAObject> uaObjects = nodeset.getUAObject();
-			if(uaObjects != null)
-			{				
-				for(UAObject object : uaObjects)
-				{
-					nodeIds.add(object.getNodeId());
-				}
-			}
-       		
-    		uaObjects = this.baseNodeset.getUAObject();
-    		ArrayList<UAObject> toRemove = new ArrayList<UAObject>();
-    		for(UAObject object : uaObjects)
-    		{
-    			if(!nodeIds.contains(object.getNodeId()))
-    			{
-    				destroyMember(object);
-    				toRemove.add(object);
-    			}
-    		}
-    		this.baseNodeset.getUAObject().removeAll(toRemove);
-    	}
-   
-    	if(this.baseNodeset.getUAVariable() != null)
-    	{	
-    		ArrayList<String> nodeIds = new ArrayList<String>();
-    		
-    		EList<UAVariable> uaVariables = nodeset.getUAVariable();
-    		if(uaVariables != null)
-    		{    			
-    			for(UAVariable var : uaVariables)
-    			{
-    				nodeIds.add(var.getNodeId());
-    			}
-    		}
-       		
-    		uaVariables = this.baseNodeset.getUAVariable();
-    		ArrayList<UAVariable> toRemove = new ArrayList<UAVariable>();
-    		for(UAVariable var : uaVariables)
-    		{
-    			if(!nodeIds.contains(var.getNodeId()))
-    			{
-    				destroyMember(var);
-    				toRemove.add(var);
-    			}
-    		}
-    		this.baseNodeset.getUAVariable().removeAll(toRemove);
-    	}  
-    	
-    	if(this.baseNodeset.getUAMethod() != null)
-    	{    		
-    		ArrayList<String> nodeIds = new ArrayList<String>();
-    		
-    		EList<UAMethod> uaMethods = nodeset.getUAMethod();
-    		if(uaMethods != null)
-    		{    			
-    			for(UAMethod method : uaMethods)
-    			{
-    				nodeIds.add(method.getNodeId());
-    			}
-    		}
-       		
-    		uaMethods = this.baseNodeset.getUAMethod();
-    		ArrayList<UAMethod> toRemove = new ArrayList<UAMethod>();
-    		for(UAMethod method : uaMethods)
-    		{
-    			if(!nodeIds.contains(method.getNodeId()))
-    			{
-    				destroyMember(method);
-    				toRemove.add(method);
-    			}
-    		}
-    		this.baseNodeset.getUAMethod().removeAll(toRemove);
-    	}
-    	
-    	if(this.baseNodeset.getUAView() != null)
-    	{    		
-    		ArrayList<String> nodeIds = new ArrayList<String>();
-
-    		EList<UAView> uaViews = nodeset.getUAView();
-    		if(uaViews != null)
-    		{    			
-    			for(UAView view : uaViews)
-    			{
-    				nodeIds.add(view.getNodeId());
-    			}
-    		}
-       		
-    		uaViews = this.baseNodeset.getUAView();
-    		ArrayList<UAView> toRemove = new ArrayList<UAView>();
-    		for(UAView view : uaViews)
-    		{
-    			if(!nodeIds.contains(view.getNodeId()))
-    			{
-    				destroyMember(view);
-    				toRemove.add(view);
-    			}
-    		}
-    		this.baseNodeset.getUAView().removeAll(toRemove);
-    	}
-		return true;
-	}
-
-	private boolean destroyMember(UANode node)
-	{
-
-		Element elem = getElement(node);
-		if(elem == null)
-		{
-			return false;
-		}
-		
-		for(Relationship rel : elem.getRelationships())
-		{
-			rel.destroy();
-		}
-		
-		this.nodeIdMap.remove(node.getNodeId());
-		this.matching.remove(elem);
-		elem.destroy();
-		
-		return true;
-	}
-
+	@SuppressWarnings("unchecked")
 	private boolean updateNamespaces(UriTable namespaceUris) {
 
 		EList<String> namespaces_new = namespaceUris.getUri();	
@@ -2904,14 +576,10 @@ public class InstanceSyncHandler {
 		if(uaElement == null)
 		{
 			uaElement = createElement(node, uaStereoType);
+			nodesToDelete.add(node); // delete old element if namespace is different
+			nodesToAdd.add(node);
 		}
-		else
-		{
-			nodesToDelete.add((UAObjectType) this.matching.get(uaElement)); // delete old element	
-			this.matching.put(uaElement, node);
-		}
-
-		nodesToAdd.add(node);
+		
 		
 		DynamicEObjectImpl stereotype = (DynamicEObjectImpl) uaElement.getStereotypeApplication(uaStereoType);
 		if(stereotype == null)
@@ -2935,14 +603,9 @@ public class InstanceSyncHandler {
 		if(uaElement == null)
 		{
 			uaElement = createElement(node, uaStereoType);
+			nodesToDelete.add(node); // delete old element if namespace is different
+			nodesToAdd.add(node);
 		}
-		else
-		{
-			nodesToDelete.add((UAVariableType) this.matching.get(uaElement)); // delete old element	
-			this.matching.put(uaElement, node);
-		}
-
-		nodesToAdd.add(node);
 		
 		
 		DynamicEObjectImpl stereotype = (DynamicEObjectImpl) uaElement.getStereotypeApplication(uaStereoType);
@@ -2967,14 +630,9 @@ public class InstanceSyncHandler {
 		if(uaElement == null)
 		{
 			uaElement = createElement(node, uaStereoType);
+			nodesToDelete.add(node); // delete old element if namespace is different
+			nodesToAdd.add(node);
 		}
-		else
-		{
-			nodesToDelete.add((UADataType) this.matching.get(uaElement)); // delete old element	
-			this.matching.put(uaElement, node);
-		}
-
-		nodesToAdd.add(node);
 		
 		DynamicEObjectImpl stereotype = (DynamicEObjectImpl) uaElement.getStereotypeApplication(uaStereoType);
 		if(stereotype == null)
@@ -3101,6 +759,7 @@ public class InstanceSyncHandler {
 		
 		if(dtd.getSymbolicName() != null)
 		{
+			@SuppressWarnings("unchecked")
 			EDataTypeUniqueEList<Object> symbollicName = (EDataTypeUniqueEList<Object>) definition.getValue(uaStereoType, "symbolicName");
 			symbollicName.clear();
 			for(String symbolic : dtd.getSymbolicName())
@@ -3122,6 +781,7 @@ public class InstanceSyncHandler {
 		Stereotype uaStereoType  = nodeSetProfile.getOwnedStereotype("DataTypeDefinition");
 		Stereotype uaDtfType  = nodeSetProfile.getOwnedStereotype("DataTypeField");
 		
+		@SuppressWarnings("unchecked")
 		EcoreEList<DynamicEObjectImpl> field = (EcoreEList<DynamicEObjectImpl>) definition.getValue(uaStereoType, "field");
 		EList<Classifier> children = definition.getNestedClassifiers();
 		
@@ -3208,6 +868,7 @@ public class InstanceSyncHandler {
 		
 		if(dtf.getDisplayName() != null)
 		{
+			@SuppressWarnings("unchecked")
 			EDataTypeUniqueEList<Object> displayName = (EDataTypeUniqueEList<Object>) field.getValue(uaStereoType, "displayName");
 			displayName.clear();
 			for(LocalizedText display : dtf.getDisplayName())
@@ -3218,6 +879,7 @@ public class InstanceSyncHandler {
 		
 		if(dtf.getDescription() != null)
 		{
+			@SuppressWarnings("unchecked")
 			EDataTypeUniqueEList<Object> descriptionList= (EDataTypeUniqueEList<Object>) field.getValue(uaStereoType, "description");
 			descriptionList.clear();
 			for(LocalizedText description : dtf.getDescription())
@@ -3241,6 +903,7 @@ public class InstanceSyncHandler {
 		
 		if(dtf.getArrayDimensions() != null && dtf.getArrayDimensions().length() > 0)
 		{
+			@SuppressWarnings("unchecked")
 			List<Object> arrayDimensions = (List<Object>) field.getValue(uaStereoType, "arrayDimensions");
 			arrayDimensions.clear();
 			
@@ -3275,6 +938,7 @@ public class InstanceSyncHandler {
 		
 		if(dtf.getSymbolicName() != null)
 		{
+			@SuppressWarnings("unchecked")
 			EDataTypeUniqueEList<Object> symbollicName = (EDataTypeUniqueEList<Object>) field.getValue(uaStereoType, "symbolicName");
 			symbollicName.clear();
 			for(String symbolic : dtf.getSymbolicName())
@@ -3297,14 +961,9 @@ public class InstanceSyncHandler {
 		if(uaElement == null)
 		{
 			uaElement = createElement(node, uaStereoType);
+			nodesToDelete.add(node); // delete old element if namespace is different
+			nodesToAdd.add(node);
 		}
-		else
-		{
-			nodesToDelete.add( (UAReferenceType) this.matching.get(uaElement)); // delete old element	
-			this.matching.put(uaElement, node);
-		}
-
-		nodesToAdd.add(node);
 		
 		DynamicEObjectImpl stereotype = (DynamicEObjectImpl) uaElement.getStereotypeApplication(uaStereoType);
 		if(stereotype == null)
@@ -3322,6 +981,7 @@ public class InstanceSyncHandler {
 			if(node.getInverseName() != null)
 			{
 				
+				@SuppressWarnings("unchecked")
 				EDataTypeUniqueEList<Object> displayNames = (EDataTypeUniqueEList<Object>) uaElement.getValue(uaStereoType, "displayName");
 				for(LocalizedText inverseName : node.getInverseName())
 				{
@@ -3346,14 +1006,9 @@ public class InstanceSyncHandler {
 		if(uaElement == null)
 		{
 			uaElement = createElement(node, uaStereoType);
+			nodesToDelete.add(node); // delete old element if namespace is different
+			nodesToAdd.add(node);
 		}
-		else
-		{
-			nodesToDelete.add((UAObject) this.matching.get(uaElement)); // delete old element	
-			this.matching.put(uaElement, node);
-		}
-
-		nodesToAdd.add(node);
 		
 		DynamicEObjectImpl stereotype = (DynamicEObjectImpl) uaElement.getStereotypeApplication(uaStereoType);
 		if(stereotype == null)
@@ -3375,6 +1030,7 @@ public class InstanceSyncHandler {
 	}
 	
 
+	@SuppressWarnings("unchecked")
 	private boolean updateOpcUAVariable(UAVariable node, List<UAVariable> nodesToAdd, List<UAVariable> nodesToDelete) {
 		Profile nodeSetProfile = this.baseUmlModel.getAppliedProfile("NodeSet");
 		Stereotype uaStereoType  = nodeSetProfile.getOwnedStereotype("UAVariable");
@@ -3384,14 +1040,9 @@ public class InstanceSyncHandler {
 		if(uaElement == null)
 		{
 			uaElement = createElement(node, uaStereoType);
+			nodesToDelete.add(node); // delete old element if namespace is different
+			nodesToAdd.add(node);
 		}
-		else
-		{
-			nodesToDelete.add((UAVariable) this.matching.get(uaElement)); // delete old element	
-			this.matching.put(uaElement, node);
-		}
-
-		nodesToAdd.add(node);
 		
 		DynamicEObjectImpl stereotype = (DynamicEObjectImpl) uaElement.getStereotypeApplication(uaStereoType);
 		if(stereotype == null)
@@ -3503,14 +1154,9 @@ public class InstanceSyncHandler {
 		if(uaElement == null)
 		{
 			uaElement = createElement(node, uaStereoType);
+			nodesToDelete.add(node); // delete old element if namespace is different
+			nodesToAdd.add(node);
 		}
-		else
-		{
-			nodesToDelete.add((UAMethod) this.matching.get(uaElement)); // delete old element	
-			this.matching.put(uaElement, node);
-		}
-
-		nodesToAdd.add(node);
 		
 		DynamicEObjectImpl stereotype = (DynamicEObjectImpl) uaElement.getStereotypeApplication(uaStereoType);
 		if(stereotype == null)
@@ -3547,6 +1193,7 @@ public class InstanceSyncHandler {
 		Stereotype uaMethodStereotype  = nodeSetProfile.getOwnedStereotype("UAMethod");
 		Stereotype uaMethodArgumentStereotype  = nodeSetProfile.getOwnedStereotype("UAMethodArgument");
 		
+		@SuppressWarnings("unchecked")
 		EcoreEList<DynamicEObjectImpl> methodArguments = (EcoreEList<DynamicEObjectImpl>) uaElement.getValue(uaMethodStereotype, "argumentDescription");
 		EList<Classifier> children = uaElement.getNestedClassifiers();
 		
@@ -3617,6 +1264,7 @@ public class InstanceSyncHandler {
 		return success;
 	}
 	
+	@SuppressWarnings("unchecked")
 	private boolean updateOpcUaMethodArgumentEntry(UAMethodArgument uaMethodArgument, Class uaElement) {
 		Profile nodeSetProfile = this.baseUmlModel.getAppliedProfile("NodeSet");
 		Stereotype uaStereoType  = nodeSetProfile.getOwnedStereotype("UAMethodArgument");
@@ -3648,14 +1296,9 @@ public class InstanceSyncHandler {
 		if(uaElement == null)
 		{
 			uaElement = createElement(node, uaStereoType);
+			nodesToDelete.add(node); // delete old element if namespace is different
+			nodesToAdd.add(node);
 		}
-		else
-		{
-			nodesToDelete.add((UAView) this.matching.get(uaElement)); // delete old element	
-			this.matching.put(uaElement, node);
-		}
-
-		nodesToAdd.add(node);
 		
 		DynamicEObjectImpl stereotype = (DynamicEObjectImpl) uaElement.getStereotypeApplication(uaStereoType);
 		if(stereotype == null)
@@ -3683,7 +1326,10 @@ public class InstanceSyncHandler {
 		return success;
 	}
 	
-	private boolean updateOpcUaNode(UANode node, Stereotype stereotype, Element uaElement) {		
+	@SuppressWarnings("unchecked")
+	private boolean updateOpcUaNode(UANode node, Stereotype stereotype, Element uaElement) {
+
+		
 		if(node.getNodeId() != null) {
 			String nodeId = node.getNodeId();
 
@@ -3762,8 +1408,6 @@ public class InstanceSyncHandler {
 			}
 
 			uaElement.setValue(stereotype, "browseName", browseName);
-			Class temp = (Class) uaElement;
-			temp.setName(browseName);
 		}
 
 	
@@ -3821,6 +1465,7 @@ public class InstanceSyncHandler {
 			Profile nodeSetProfile = this.baseUmlModel.getAppliedProfile("NodeSet");
 			Stereotype uaNodeSetType  = nodeSetProfile.getOwnedStereotype("UANodeSetType");
 			
+			@SuppressWarnings("unchecked")
 			EcoreEList<DynamicEObjectImpl> nsList = (EcoreEList<DynamicEObjectImpl>) this.baseUmlModel.getValue(uaNodeSetType, "nameSpaceUris");
 			
 			if(nsList.size() < namespaceId)
@@ -3900,6 +1545,7 @@ public class InstanceSyncHandler {
 			Profile nodeSetProfile = this.baseUmlModel.getAppliedProfile("NodeSet");
 			Stereotype uaNodeSetType  = nodeSetProfile.getOwnedStereotype("UANodeSetType");
 			
+			@SuppressWarnings("unchecked")
 			EcoreEList<DynamicEObjectImpl> nsList = (EcoreEList<DynamicEObjectImpl>) this.baseUmlModel.getValue(uaNodeSetType, "nameSpaceUris");
 			
 			if(nsList.size() < namespaceId)
@@ -4313,6 +1959,7 @@ public class InstanceSyncHandler {
 				Profile nodeSetProfile = this.baseUmlModel.getAppliedProfile("NodeSet");
 				Stereotype uaNodeSet  = nodeSetProfile.getOwnedStereotype("UANodeSetType");
 				
+				@SuppressWarnings("unchecked")
 				EcoreEList<DynamicEObjectImpl> nsList = (EcoreEList<DynamicEObjectImpl>) this.baseUmlModel.getValue(uaNodeSet, "nameSpaceUris");
 				int seperator = nodeId.indexOf(";");
 				String nameSpaceIdString = nodeId.substring(3, seperator); // ns= is 3 chars long
@@ -4488,6 +2135,7 @@ public class InstanceSyncHandler {
 		return success;
 	}
 	
+	@SuppressWarnings("unchecked")
 	private boolean updateOpcUaRolePermission(UANode rolePermissionNode) {
 		
 		Class parent = (Class) getElement(rolePermissionNode);
@@ -4695,13 +2343,7 @@ public class InstanceSyncHandler {
 			if(name2.equalsIgnoreCase("base_Class"))
 			{
 				baseClass= (Class)eObject.dynamicGet(id2);
-								
-				if(!this.matching.containsKey(baseClass) && addIfNotExists &&
-					this.baseUmlModel.equals(baseClass.getModel()))
-				{
-					transformClass(baseClass);
-				}
-
+				
 				break;
 			}
 		}
@@ -4716,19 +2358,7 @@ public class InstanceSyncHandler {
 		Package namespace = umlUaNode.getNearestPackage();
 		Stereotype uaStereotype;
 		
-		if(this.baseUmlModel.equals(umlUaNode.getModel()))
-		{				
-			if(!this.matching.containsKey(umlUaNode))
-			{				
-				transformClass(umlUaNode);
-			}
-			uaStereotype = getMatchingStereotype(this.matching.get(umlUaNode));
-		}
-		else
-		{
-			// the other method is faster, therefore it should be used by default
-			uaStereotype = getMatchingStereotype(umlUaNode);
-		}
+		uaStereotype = getMatchingStereotype(umlUaNode);
 		
 		if(umlUaNode.hasValue(uaStereotype, "nodeId"))
 		{
@@ -4770,3 +2400,4 @@ public class InstanceSyncHandler {
 	}
 
 }
+
