+++ at.ac.tuwien.auto.modelua.papyrus.opcua.diagram.transformation/src/at/ac/tuwien/auto/modelua/papyrus/opcua/diagram/transformation/umltoopcua/TransformUMLPackageCommand.java
@@ -1,20 +1,20 @@
-package at.ac.tuwien.auto.modelua.papyrus.opcua.diagram.transformation.transformer;
+package at.ac.tuwien.auto.modelua.papyrus.opcua.diagram.transformation.umltoopcua;
 
 import org.eclipse.emf.transaction.RecordingCommand;
 import org.eclipse.emf.transaction.TransactionalEditingDomain;
 import org.eclipse.uml2.uml.Package;
 
-public class UpdateUMLPackageCommand extends RecordingCommand {
+public class TransformUMLPackageCommand extends RecordingCommand {
 
-	private InstanceSyncHandler handler;
+	private UmlToOpcUaTransformer handler;
 	private Package obj;
 	private boolean success = false;
 	
-	public UpdateUMLPackageCommand(TransactionalEditingDomain domain) {
+	public TransformUMLPackageCommand(TransactionalEditingDomain domain) {
 		super(domain);
 	}
 
-	public void registerInstanceSynchHandler(InstanceSyncHandler handler)
+	public void registerInstanceSynchHandler(UmlToOpcUaTransformer handler)
 	{
 		this.handler = handler;
 	}
